Extract TagCanonicalizer
